almost finish the filter
