+++ ASAC/about.html
@@ -10,15 +10,15 @@
     <meta content="" name="keywords">
 
     <!-- Favicons -->
-    <link href="assets/images/asac_ltuc_logo.jpg" rel="icon">
+    <link href="../assets/images/asac_ltuc_logo.jpg" rel="icon">
 
     <!-- Template Main CSS File -->
-    <link href="../reset_default.css" rel="stylesheet">
+    <link href="../assets/css/reset_default.css" rel="stylesheet">
     <script src="https://kit.fontawesome.com/1f66153deb.js" crossorigin="anonymous"></script>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css" />
-    <link href="../style.css" rel="stylesheet">
+    <link href="../assets/css/style.css" rel="stylesheet">
     <style>
         body {
             background-color: #f4f4f4;
@@ -145,19 +145,13 @@
                             <a class="nav-link" href="programs.html"> Programs </a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="../contactUs.html">Contact Us</a>
+                            <a class="nav-link" href="contactUs.html">Contact Us</a>
                         </li>
 
                     </ul>
                 </div>
             </div>
         </nav>
-        <!-- <svg class='burger-menu-icon' width="24" height="16" viewBox="0 0 24 16" fill="none"
-                xmlns="http://www.w3.org/2000/svg">
-                <path
-                    d="M1.33334 16C0.95556 16 0.638671 15.872 0.382671 15.616C0.126671 15.36 -0.000884275 15.0436 4.61361e-06 14.6667C4.61361e-06 14.2889 0.128005 13.972 0.384005 13.716C0.640005 13.46 0.956449 13.3324 1.33334 13.3333H22.6667C23.0444 13.3333 23.3613 13.4613 23.6173 13.7173C23.8733 13.9733 24.0009 14.2898 24 14.6667C24 15.0444 23.872 15.3613 23.616 15.6173C23.36 15.8733 23.0436 16.0009 22.6667 16H1.33334ZM1.33334 9.33334C0.95556 9.33334 0.638671 9.20534 0.382671 8.94934C0.126671 8.69334 -0.000884275 8.37689 4.61361e-06 8C4.61361e-06 7.62223 0.128005 7.30534 0.384005 7.04934C0.640005 6.79334 0.956449 6.66578 1.33334 6.66667H22.6667C23.0444 6.66667 23.3613 6.79467 23.6173 7.05067C23.8733 7.30667 24.0009 7.62312 24 8C24 8.37778 23.872 8.69467 23.616 8.95067C23.36 9.20667 23.0436 9.33423 22.6667 9.33334H1.33334ZM1.33334 2.66667C0.95556 2.66667 0.638671 2.53867 0.382671 2.28267C0.126671 2.02667 -0.000884275 1.71023 4.61361e-06 1.33334C4.61361e-06 0.95556 0.128005 0.638672 0.384005 0.382672C0.640005 0.126672 0.956449 -0.000884275 1.33334 4.6136e-06H22.6667C23.0444 4.6136e-06 23.3613 0.128005 23.6173 0.384005C23.8733 0.640005 24.0009 0.956449 24 1.33334C24 1.71112 23.872 2.028 23.616 2.284C23.36 2.54 23.0436 2.66756 22.6667 2.66667H1.33334Z"
-                    fill="white" />
-            </svg> -->
         </div>
 
     </header>
@@ -166,7 +160,7 @@
 
         <section id="banner" class="prog-header"
             style="height: 250px;background-image: linear-gradient(to top, rgb(22 21 21 / 60%), rgb(22 21 21 / 60%)),url('https://placehold.co/800?text=placeholder');">
-            <h1 class="fs-1">
+            <h1 class="fs-1 animate__animated animate__zoomIn">
                 About Us
             </h1>
 
@@ -337,12 +331,7 @@
                                 Feedback and question
                             </li>
                         </a>
-                        <a href="None">
-                            <li>
-                                News
-                            </li>
-                        </a>
-                        <a href="../index.html#contact">
+                        <a href="contactUs.html">
                             <li>
                                 Contact us
                             </li>
@@ -396,7 +385,7 @@
         </div>
         <div class="footer-lower">
 
-            <a href="index.html">
+            <a href="../index.html">
                 <img src="../assets/images/LTUCLogo.png" class="logo footer" loading="lazy">
             </a>
 
@@ -447,7 +436,7 @@
         </div>
     </footer>
 
-    <script src="../app.js"></script>
+    <script src="../assets/js/app.js"></script>
     <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/@emailjs/browser@3/dist/email.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
