+++ ASAC/contactUs.html
@@ -11,15 +11,15 @@
     <meta content="" name="keywords">
 
     <!-- Favicons -->
-    <link href="assets/images/asac_ltuc_logo.jpg" rel="icon">
+    <link href="../assets/images/asac_ltuc_logo.jpg" rel="icon">
 
     <!-- Template Main CSS File -->
-    <link href="reset_default.css" rel="stylesheet">
+    <link href="../assets/css/reset_default.css" rel="stylesheet">
     <script src="https://kit.fontawesome.com/1f66153deb.js" crossorigin="anonymous"></script>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css" />
-    <link href="style.css" rel="stylesheet">
+    <link href="../assets/css/style.css" rel="stylesheet">
 
 </head>
 <!-- ************** CONTACT *************** -->
@@ -28,8 +28,8 @@
     <header>
         <nav class="navbar navbar-expand-lg ">
             <div class="container-fluid">
-                <a class="navbar-brand logo" href="index.html">
-                    <img src="assets/images/asac-logo.png" width="200" alt="Logo" loading="lazy">
+                <a class="navbar-brand logo" href="../index.html">
+                    <img src="../assets/images/asac-logo.png" width="200" alt="Logo" loading="lazy">
                 </a>
                 <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
                     data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
@@ -39,13 +39,13 @@
                 <div class="collapse justify-content-end navbar-collapse" id="navbarSupportedContent">
                     <ul class="navbar-nav mb-2 mb-lg-0 align-items-center">
                         <li class="nav-item">
-                            <a class="nav-link" href="index.html">Home</a>
+                            <a class="nav-link" href="../index.html">Home</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="ASAC/about.html">About</a>
+                            <a class="nav-link" href="about.html">About</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="ASAC/programs.html"> Programs </a>
+                            <a class="nav-link" href="programs.html"> Programs </a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link" href="contactUs.html">Contact Us</a>
@@ -63,7 +63,7 @@
 
         <section id="banner" class="prog-header"
             style="height: 250px;background-image: linear-gradient(to top, rgb(22 21 21 / 60%), rgb(22 21 21 / 60%)),url('https://placehold.co/800?text=placeholder');">
-            <h1 class="fs-1">
+            <h1 class="fs-1 animate__animated animate__zoomIn">
                 Contact Us
             </h1>
 
@@ -241,12 +241,7 @@
                                 Feedback and question
                             </li>
                         </a>
-                        <a href="None">
-                            <li>
-                                News
-                            </li>
-                        </a>
-                        <a href="../index.html#contact">
+                        <a href="contactUs.html">
                             <li>
                                 Contact us
                             </li>
@@ -300,8 +295,8 @@
         </div>
         <div class="footer-lower">
 
-            <a href="index.html">
-                <img src="./assets/images/LTUCLogo.png" class="logo footer" loading="lazy">
+            <a href="../index.html">
+                <img src="../assets/images/LTUCLogo.png" class="logo footer" loading="lazy">
             </a>
 
             <div class="copyrights">
@@ -351,7 +346,7 @@
         </div>
     </footer>
 
-    <script src="app.js"></script>
+    <script src="../assets/js/app.js"></script>
     <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/@emailjs/browser@3/dist/email.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
@@ -375,11 +370,9 @@
         let alertContent = document.getElementById('alertContent');
         let category = document.getElementById('q-category').value;
         let recipientEmail = getRecipientEmail(category);
-        console.log('***** recipientEmail', recipientEmail);
 
         (function () {
             // https://dashboard.emailjs.com/admin/account
-            console.log('publicKey', recipientEmail.publicKey);
             emailjs.init(recipientEmail.publicKey);
         })();
         let templateParams = {
@@ -440,8 +433,8 @@
                 'employment': {
                     email: 'employment.asac@ltuc.com',
                     serviceID: '',
-                    templateID: '',
-                    publicKey: ''
+                    templateID: 'template_0l2zsoc',
+                    publicKey: 'Lw1TDWJ761Q3KfwdH'
                 },
             };
             return query_categories[category];
