+++ assets/css/style.css
@@ -34,7 +34,7 @@ h3 {
   font-size: 1em !important;
 }
 
-p {
+p, option {
   font-size: 0.9em !important;
 }
 
@@ -98,7 +98,34 @@ a {
   align-items: center;
   gap: 15px;
 }
-
+.prog-result {
+  color: #962F8A;
+  margin-top: 20px;
+}
+.prog-icon{
+  width: 50px;
+  height: 50px;
+  
+}
+.gray-color{
+  color: gray;
+}
+.btn-primary {
+  --bs-btn-color: #fff;
+  --bs-btn-bg: #962F8A;
+  --bs-btn-border-color: #962F8A;
+  --bs-btn-hover-color: #fff;
+  --bs-btn-hover-bg: rgb(130, 31, 118);
+  --bs-btn-hover-border-color: rgb(130, 31, 118);
+  --bs-btn-focus-shadow-rgb: 49, 132, 253;
+  --bs-btn-active-color: #fff;
+  --bs-btn-active-bg: rgb(130, 31, 118);
+  --bs-btn-active-border-color: rgb(130, 31, 118);
+  --bs-btn-active-shadow: inset 0 3px 5px rgba(0, 0, 0, 0.125);
+  --bs-btn-disabled-color: #fff;
+  --bs-btn-disabled-bg: #962F8A;
+  --bs-btn-disabled-border-color: #962F8A;
+}
 .code-sign svg {
   width: 35px;
 }
