+++ index.html
@@ -13,12 +13,12 @@
     <link href="assets/images/asac_ltuc_logo.jpg" rel="icon">
 
     <!-- Template Main CSS File -->
-    <link href="reset_default.css" rel="stylesheet">
+    <link href="assets/css/reset_default.css" rel="stylesheet">
     <script src="https://kit.fontawesome.com/1f66153deb.js" crossorigin="anonymous"></script>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css" />
-    <link href="style.css" rel="stylesheet">
+    <link href="assets/css/style.css" rel="stylesheet">
 
 </head>
 
@@ -47,7 +47,7 @@
                             <a class="nav-link" href="ASAC/programs.html"> Programs </a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="contactUs.html">Contact Us</a>
+                            <a class="nav-link" href="ASAC/contactUs.html">Contact Us</a>
                         </li>
 
                     </ul>
@@ -69,8 +69,6 @@
             <div class="banner-content">
                 <div class="banner-content-inner">
                     <h1 class="animate__animated animate__fadeInDown">Abdul Aziz Al Ghurair School of Advanced Computing</h1>
-                    <!-- <p>Welcome to the Abdul Aziz Al Ghurair School of Advanced Computing, a hub of innovation and
-                        excellence in technology education</p> -->
                     <div class="banner-btns">
                         <a class="animate__animated animate__fadeInLeft" href="ASAC/programs.html">Explore Our Programs</a>
                         <a class="animate__animated animate__fadeInRight" href="#">Be a part of our Family</a>
@@ -173,7 +171,7 @@
                 <p>Experience innovation at our campus! Join a vibrant community, cutting-edge facilities, and a
                     supportive environment. Elevate your education with us – where learning meets inspiration</p>
             </div>
-            <div class='carosel left'>
+            <div class='carosel center'>
                 <div class='scroll-img'>
                     <div class='carosel-img'>
                         <img src='https://placehold.co/800?text=placeholder' loading="lazy">
@@ -213,7 +211,6 @@
 
 
         <!-- Program  -->
-
         <section id="programs-brief " class="container	">
             <div class='ltuc-description '>
                 <h1>Programs</h1>
@@ -320,7 +317,6 @@
 
 
         <!-- Sucess Story  -->
-
         <section id="success-breif " class="container	">
             <div class='ltuc-description '>
                 <h1>Sucess Stories</h1>
@@ -403,18 +399,7 @@
             <div class="news-cards">
 
             </div>
-            <!-- <button type="button" class=" latest-see-more">See More</button> -->
         </section>
-
-
-        <!-- <div class="hide-on-desktop go-top">
-            <a href="#top">
-                <svg xmlns="http://www.w3.org/2000/svg" height="16" width="12" viewBox="0 0 384 512">
-                    <path opacity="1" fill="#ffffff"
-                        d="M214.6 41.4c-12.5-12.5-32.8-12.5-45.3 0l-160 160c-12.5 12.5-12.5 32.8 0 45.3s32.8 12.5 45.3 0L160 141.2V448c0 17.7 14.3 32 32 32s32-14.3 32-32V141.2L329.4 246.6c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3l-160-160z" />
-                </svg>
-            </a>
-        </div> -->
     </main>
 
     <!-- ************** FOOTER *************** -->
@@ -426,17 +411,12 @@
 
                     <h3>About</h3>
                     <ul>
-                        <a href="FAQ.html">
+                        <a href="ASAC/FAQ.html">
                             <li>
                                 Feedback and question
                             </li>
                         </a>
-                        <a href="None">
-                            <li>
-                                News
-                            </li>
-                        </a>
-                        <a href="../index.html#contact">
+                        <a href="ASAC/contactUs.html">
                             <li>
                                 Contact us
                             </li>
@@ -541,8 +521,7 @@
         </div>
     </footer>
 
-    <script src="app.js"></script>
-    <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/@emailjs/browser@3/dist/email.min.js"></script>
+    <script src="assets/js/app.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
         crossorigin="anonymous"></script>
